backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
python-dotenv
groq
langchain
langchain-groq
langchain-community
langchain-chroma
chromadb
fastembed